dashboard: Add basic query and query result types
This sets the skeleton for the Dashboard library where we will build
infrastructure to build dashboard backends for various projects.

The commit adds a basic query type to represent a query for data needed
to display in a single dashboard component, along with a result type to
represent corresponding results.

We use tuples and arrays of tuples in a number of places, and these
might be replaced by type constructors and maps respectively in the
future.