dashboard: Make QueryResult type more like database rows
The QueryResult was previously being constructed based on the query
groups. While this sort of construction will likely need to be done,
it's best left to the frontend which knows what shape it wants the data
in best.

So we now model the output more or less as database rows (with the time
intervals being explicitly parsed out to be symmetrical with the Query).