console: Make test more robust and not dependent on external data
We now include the required BigQuery credentials in the test fixtures,
and use a public dataset which should be reliable in terms of the result
not changing (as we were facing with internal data that was getting
updated, presumably as refunds etc. were backfilled).

This also means that we can (and in this commit, do) enable tests in the
nix environment, and thus in CI.