console: Add a way to transform query result strings
This adds a mechanism to transform string values from database queries
into human readable strings. This is currently implemented as a manual
mapping of each expected value and the corresponding transformation. At
some point we may also want to add the ability to specify transformation
functions (for example, capitalise-first-letter-only).

The implementation is that we transform the QueryResultRow from the
database before sending it to the client. The mapping of names is
provided in configuration. This does make the config file more verbose,
but this seems okay for the moment.

Because of the version of Dhall that we are using, the "map" is actually
a [(String, String)]. Newer versions support reading this sort of
structure into a (Map String String). Since this is likely not a huge
concern right now, I'm choosing to let the code be less verbose rather
than more efficient. We can switch the a newer Dhall version and use
this Map functionality in the future easily enough.