dashboard: Support arbitrary dates in backend
We currently have a hard coded date in the Console backend to avoid scanning
too much data in queries. The original plan was to move to Vision tables and
avoid this issue.

As we currently cannot use Vision data, we will need to use BQ's wildcard and
table suffix features to avoid arbitrary date queries in the backend.

Implement the table suffix feature so we can support this functionality.

For the console test we now have a separate table of our own just for
the tests. This table is reflected in the tests now and is in the dataset
bq-dashboard-test. The table is sharded by date and has dummy merchant
ids and random data for rest of fields. Reason for moving from the
instant-text dataset used earlier, is that dataset was not sharded and
we would not have able to test our _TABLE_SUFFIX implementation.

MUSA-119

Signed-off-by: Sanchayan Maity <sanchayan.maity@juspay.in>