console: Switch to using a custom monad stack with Servant
Till now we were using Servant's default monad transformer stack which
is simplistic. Add a Reader monad on top of Servant's default Handler
which should also allow us to not have to thread the query backend and
query configuration through each function.

This will also act as the groundwork for adding logging related
configuration and context in the configuration we carry around with
Reader.

Signed-off-by: Sanchayan Maity <sanchayan.maity@juspay.in>