Use unagi-chan for logging queue
Problem: using STM in cases with multiple producers and consumers is
dangerous for performance, better use some specific IO-based queue.

Also, the queue should be bounded if we want to avoid unlimited memory
consumption - blocking the caller is justified in case when queue is
overflowed.

Solution: unagi-chan seems to be the most advised solution here, let's
use bounded version of the queue from here.