Annotate the necessary types with kinds
Problem: `DBConfig` and some other types have a phantom type parameter
`beM` which is of arbitrary kind `k`, this may give problems if user has
`PolyKinds` extension disabled as any helpers over `DBConfig beM` will
have `beM` default to kind `Type`.

Solution: since `Pg`, `Mysql` and other instantiations for `beM` are of
`Type -> Type` kind, we constraint the kind of the entire `beM` to that.

Also I disable `PolyKinds` extension - this way it is immediately
obvious when some type parameter is of wrong type, and this actually
revealed several other types for which we should specify the kind of
type arguments.
Normally it is not adviced to use `PolyKinds` in the code related to
business logic for the mentioned reason and some other ones. It can be
enabled to provide a better control in modules which work with advanced
type-level stuff, but with more control comes the greater
responsibility.