Nix: make remoteDeps work in nix-shell
- Replace the roundabout way (mk-shell.nix) with direct invocation of
`shellFor`, because it is much simpler and easier to reason about. It
also fixes the bug for top-level nix-shell. I would recommend moving
away from using mk-shell.nix, and eventually retiring it.
- Remove shell.nix, because in its absense default.nix is automatically
used, and default.nix checks if nix-shell or nix-build is invoked using
`inNixShell`.

In summary, use one Nix file -- `default.nix` -- and do shell-specific
stuff in it.